enable parse paging

+++ DayTraderz/ParseClient.h
@@ -15,8 +15,8 @@
 
 + (ParseClient *)instance;
 - (void)fetchAccountForUser:(PFUser *)user callback:(void(^)(NSArray *objects, NSError *error))callback;
-- (void)fetchPicksForAccount:(Account *)account withSkip:(int)skip callback:(void(^)(NSArray *objects, NSError *error))callback;
-- (void)fetchLeadersSortedByColumn:(NSString *)column withSkip:(int)skip callback:(void(^)(NSArray *objects, NSError *error))callback;
+- (void)fetchPicksForAccount:(Account *)account withSkip:(long)skip callback:(void(^)(NSArray *objects, NSError *error))callback;
+- (void)fetchLeadersSortedByColumn:(NSString *)column withSkip:(long)skip callback:(void(^)(NSArray *objects, NSError *error))callback;
 - (void)createOrUpdatePick:(Pick *)pick;
 
 @end
